+++ src/main/java/lotto/Application.java
@@ -5,7 +5,7 @@ import controller.LottoController;
 public class Application {
     public static void main(String[] args) {
         LottoController lottoController = new LottoController();
-        
+
         lottoController.gameStart();
     }
 }
